Add basic test of template parsing
There's not yet a great way to test for all possible ways a template
could be invalid. At the very least we can gate against invalid syntax.
This patch adds a test to ensure that we continue to fail component
creation when at least one of the template strings are invalid.