Fix interpretation of Content-Length header
The previous version checked for content-length as a gate before
attempting to read from the response body stream. However, not all
servers report content-length correctly and automatically. In these
cases the interpretation of the response body was skipped by the
content-length check. This change adjusts the switch to check the length
of the content read from the body rather than the content-length header.