Add AWS request signing for Lambda calls
This adds auth signature headers to outgoing requests using the
AWS SDK. These signatures are required to authenticate with a Lambda
endpoint.

A big note here is that it will now be required to pass in a valid AWS
region and either have access to valid AWS credentials or pass in fake
static credentials to trick the signature process. The benefit of this
change is that real AWS Lambda endpoints are now valid targets where
before calls would fail due to auth.